Fix issue where when rule fix was adding in new lines, the lines would start with 2 space characters, new lines are now completely empty
