+++ index.js
@@ -1,5 +1,5 @@
 // Unopinionated extensible tslint configuration
 // Loads rules for extending packages, but does not enable any
 module.exports = {
-  rulesDirectory: './'
+  rulesDirectory: '.'
 };
