+++ package.json
@@ -6,7 +6,7 @@
     "compile": "tsc",
     "prepublish": "npm run compile",
     "lint": "tslint --config tslint.json --exclude node_modules **/*.ts",
-    "test": "ava-ts src/**/*.spec.ts --verbose",
+    "test": "ava-ts src/**/test/**/*.spec.ts --verbose",
     "test:watch": "npm test -- --watch"
   },
   "keywords": [
@@ -26,7 +26,7 @@
     "ava-ts": "^0.24.3",
     "temp-dir": "^1.0.0",
     "ts-node": "^6.0.0",
-    "tslint": "^5.9.1",
-    "typescript": "2.3.4"
+    "typescript": "3.4.5",
+    "tslint": "5.15.0"
   }
 }
