+++ test/fixtures/fails/fixes/noLineAndCommentAboveClass.ts
@@ -1,6 +1,6 @@
 // this is a comment
 class NoLineAndCommentAboveClass {
   constructor() {}
-  
+
   method() {}
 }
