+++ test/fixtures/fails/fixes/noLineAndCommentBetweenVarAndConstructor.ts
@@ -1,7 +1,7 @@
 class NoLineAndCommentBetweenVarAndConstructor {
 
   myVar: string = 'my nice string';
-  
+
   // this is a constructor
   constructor() {}
