+++ test/fixtures/fails/fixes/noLineAndCommentBetweenVarAndMethod.ts
@@ -3,14 +3,14 @@ class NoLineAndCommentBetweenVarAndMethod {
   constructor() {}
 
   myVar: string = 'my nice string';
-  
+
   /**
    * Documenting this method with block comment
    */
   method() {}
 
   myNum: number = 24;
-  
+
   // Documenting this method with single line comment
   method() {}
 }
