+++ test/fixtures/fails/fixes/noLineAndMethodComment.ts
@@ -1,12 +1,12 @@
 class NoLineAndMethodComment {
 
   constructor() {}
-  
+
   /**
    * Documenting this method with block comment
    */
   method() {}
-  
+
   // Documenting this method with single line comment
   anotherMethod() {}
 }
