+++ test/fixtures/fails/fixes/noLineBetweenVarAndConstructor.ts
@@ -1,7 +1,7 @@
 class NoLineBetweenVarAndConstructor {
 
   myVar: string = 'my nice string';
-  
+
   constructor() {}
 
   method() {}
