+++ test/fixtures/fails/fixes/noLineBetweenVarAndMethod.ts
@@ -3,6 +3,6 @@ class NoLineBetweenVarAndMethod {
   constructor() {}
 
   myVar: string = 'my nice string';
-  
+
   method() {}
 }
