+++ test/fixtures/fails/fixes/oneLineNoNewLines.ts
@@ -1,5 +1,5 @@
 class NoNewLines {
   constructor() {}
-  
+
   method() {}
 }
