+++ test/fixtures/fails/fixes/oneLineTwoNewLines.ts
@@ -1,5 +1,5 @@
 class TwoNewLines {
-  
+
   constructor() {}
 
   method() {}
