+++ test/fixtures/fails/fixes/twoLinesOneNewLine.ts
@@ -1,8 +1,8 @@
 class OneNewLine {
-  
-  
+
+
   constructor() {}
 
-  
+
   method() {}
 }
