+++ test/fixtures/fails/fixes/twoLinesThreeNewLines.ts
@@ -1,5 +1,5 @@
 class ThreeNewLines {
-  
+
 
   constructor() {}
