+++ test/fixtures/fails/oneLineTwoNewLines.ts
@@ -1,6 +1,6 @@
 class TwoNewLines {
-  
-  
+
+
   constructor() {}
