+++ test/fixtures/fails/twoLinesOneNewLine.ts
@@ -1,5 +1,5 @@
 class OneNewLine {
-  
+
   constructor() {}
 
   method() {}
