+++ test/fixtures/fails/twoLinesThreeNewLines.ts
@@ -1,7 +1,7 @@
 class ThreeNewLines {
-  
 
-  
+
+
   constructor() {}
