+++ tsconfig.json
@@ -14,6 +14,6 @@
     "src"
   ],
   "exclude": [
-    "**/*.spec.ts"
+    "src/test/**/*"
   ]
 }
